# packages/backend/openapi/openapi.yaml
...
openapi: '3.1.0'
info:
  title: Linked Data Explorer Backend
  description: >-
    DMN discovery and chain execution, SHACL validation, asset storage and
    integrations for the Linked Data Explorer.

    Every 4xx/5xx response is `application/problem+json` (RFC 9457,
    schema `Problem`): `status`, `title` and `detail` always; `code`, the
    machine-readable identifier existing callers branch on, where the
    handler sets one. Two operations are deliberate exceptions, each noted
    at its own response: `POST /dmns/evaluate/{decisionKey}` is a byte-for-byte
    proxy for Operaton's own evaluate endpoint and passes Operaton's error
    body through unchanged (or answers `OperatonProxyError` when Operaton
    could not be reached at all), and `GET /health`'s 503 body is the health
    document itself, not a problem response, because the document *is* the
    explanation. https://github.com/sgort/linked-data-explorer/issues/131.
  contact:
    name: Steven Gort
    email: steven.gort@ictu.nl
    url: https://iou-architectuur.open-regels.nl/
servers:
  - url: https://backend.linkeddata.open-regels.nl/v1
    description: Production
  - url: https://acc.backend.linkeddata.open-regels.nl/v1
    description: Acceptance
...
paths:
...
  /assets/ropa:
    get:
      operationId: listRopa
      tags: [Assets]
      summary: List stored records of processing activities
      description: >-
        Every row the table holds, most recently updated first — draft and
        archived records included, not only active ones (contrast
        `GET /ropa/public`, which filters to `status = active`).
      responses:
        '200':
          description: The stored RoPA records, most recently updated first.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RopaRecord'
        '500':
          description: The read failed (code LIST_FAILED).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: >-
            The database pool is not configured (code DB_NOT_CONFIGURED).
            Every `/assets/ropa` operation can answer this.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: upsertRopa
      tags: [Assets]
      summary: Create or update a record of processing activities
      description: >-
        Upserts by `bpmnProcessId`: inserts a new row, or on a conflict with
        an existing `bpmnProcessId` updates every column named below and
        replaces the full set of personal-data fields (deleted and
        re-inserted inside the same transaction). `id`,
        `createdAt` and `updatedAt` are never read from the body — the id
        is database-generated and both timestamps are set by the database
        clock on insert and update, unlike `/assets/forms` and
        `/assets/documents`, where the caller supplies them. Several
        columns have a database default (`thirdCountryTransfers` false,
        `status` draft, `schemaVersion` 1, each personal-data field's
        `specialCategory` false and `sortOrder` 0), but this operation
        always supplies a value for each explicitly, even when the caller
        omits it from the body — an omitted value becomes the JavaScript
        value `undefined`, which the Postgres driver sends as SQL NULL,
        and NULL is rejected by each column's NOT NULL constraint. The
        database default therefore never actually applies through this
        route. Nothing in the body is validated: a value a database CHECK
        constraint rejects (`processLevel`, `status`), a required column
        left empty, or an omitted `personalDataFields` array (iterated
        directly, so a missing array throws before any database call),
        answers 500, never 400.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >-
                Passed straight to the database with no validation; the
                fields below are what a successful upsert needs, not what
                is enforced before that point.
              required:
                - bpmnProcessId
                - processLevel
                - title
                - controllerName
                - controllerContact
                - purpose
                - legalBasisUri
                - legalBasisLabel
                - gdprArticle
                - dataSubjects
                - recipients
                - thirdCountryTransfers
                - retentionPeriod
                - securityMeasures
                - status
                - schemaVersion
                - personalDataFields
              properties:
                bpmnProcessId:
                  type: string
                processLevel:
                  type: string
                  enum: [shell, subprocess]
                  description: Database CHECK constraint; an out-of-set value answers 500.
                title:
                  type: string
                controllerName:
                  type: string
                controllerContact:
                  type: string
                dpoContact:
                  type: string
                purpose:
                  type: string
                legalBasisUri:
                  type: string
                legalBasisLabel:
                  type: string
                gdprArticle:
                  type: string
                dataSubjects:
                  type: string
                recipients:
                  type: string
                thirdCountryTransfers:
                  type: boolean
                thirdCountryDetails:
                  type: string
                retentionPeriod:
                  type: string
                securityMeasures:
                  type: string
                status:
                  type: string
                  enum: [draft, active, archived]
                  description: Database CHECK constraint; an out-of-set value answers 500.
                schemaVersion:
                  type: integer
                personalDataFields:
                  type: array
                  description: >-
                    Replaces the record's full set of personal-data fields.
                    An empty array is valid and deletes any fields the
                    record already had.
                  items:
                    type: object
                    description: >-
                      Distinct from the response's RopaPersonalDataField:
                      no `id` or `ropaRecordId` here — both are
                      database-generated on insert.
                    required:
                      [formId, fieldKey, fieldLabel, dataCategory, specialCategory, sortOrder]
                    properties:
                      formId:
                        type: string
                      fieldKey:
                        type: string
                      fieldLabel:
                        type: string
                      dataCategory:
                        type: string
                        description: No database CHECK constraint; any string is accepted.
                      specialCategory:
                        type: boolean
                      sortOrder:
                        type: integer
      responses:
        '200':
          description: >-
            The record was created or updated. The response carries only
            the record's id, not the full record.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    additionalProperties: false
                    required: [id]
                    properties:
                      id:
                        type: string
                        format: uuid
        '500':
          description: >-
            The write failed (code UPSERT_FAILED) — for example a database
            CHECK or NOT NULL violation, or a missing `personalDataFields`
            array — or the request body was not valid JSON or over 10 MB
            (code INTERNAL_ERROR; #143).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: >-
            The database pool is not configured (code DB_NOT_CONFIGURED).
            Every `/assets/ropa` operation can answer this.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /assets/ropa/by-bpmn-id/{bpmnProcessId}:
    get:
      operationId: getRopaByBpmnProcessId
      tags: [Assets]
      summary: Resolve a RoPA record by its BPMN process id
      description: >-
        Looks up the most recently updated record for that BPMN process id.
      parameters:
        - name: bpmnProcessId
          in: path
          required: true
          description: >-
            The BPMN process id to resolve. Any string is accepted; a value
            matching no row answers 404, never 400.
          schema:
            type: string
      responses:
        '200':
          description: The record was found.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/RopaRecord'
        '404':
          description: >-
            No RoPA record has this bpmnProcessId (code NOT_FOUND). The
            message quotes the requested id verbatim.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: The lookup failed (code LOOKUP_FAILED).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: >-
            The database pool is not configured (code DB_NOT_CONFIGURED).
            Every `/assets/ropa` operation can answer this.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /assets/ropa/{id}:
    delete:
      operationId: deleteRopa
      tags: [Assets]
      summary: Delete a stored RoPA record
      description: >-
        Deletes by `id`, cascading to its personal-data fields. The delete
        statement is unconditional — an id matching no row still answers
        200, never 404. `id` is a database UUID column: a non-UUID-shaped
        value fails Postgres's implicit cast and answers 500, not 400 or
        404.
      parameters:
        - name: id
          in: path
          required: true
          description: The record id to delete.
          schema:
            type: string
      responses:
        '200':
          description: >-
            The delete ran. The response carries no data and does not say
            whether a row was actually removed.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    const: true
        '500':
          description: >-
            The delete failed (code DELETE_FAILED) — including a
            non-UUID-shaped id, which fails the column's implicit cast.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: >-
            The database pool is not configured (code DB_NOT_CONFIGURED).
            Every `/assets/ropa` operation can answer this.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  headers:
    ApiVersion:
      description: Full release version of the API (/core/version-header).
      required: true
      schema:
        type: string
        examples:
          - '2026.09.4'
  schemas:
    Problem:
      type: object
      description: >-
        RFC 9457 problem details (https://www.rfc-editor.org/rfc/rfc9457),
        returned as `application/problem+json` by every 4xx/5xx response
        except two: `POST /dmns/evaluate/{decisionKey}` forwards Operaton's
        own error body, and `GET /health`'s 503 is the health document
        itself -- both documented at their own operation. `code` is kept as
        an extension member alongside the RFC's own fields: the
        machine-readable identifier existing callers already branch on.
        https://github.com/sgort/linked-data-explorer/issues/131.
      required: [status, title, detail]
      properties:
        type:
          type: string
          format: uri-reference
          description: >-
            A URI identifying the problem type. `about:blank` (RFC 9457
            §4.2.1) unless an operation's description says otherwise --
            every problem is that operation's own generic 4xx/5xx, not a
            documented category with its own reference page.
          examples: [about:blank]
        status:
          type: integer
          description: The HTTP status, repeated from the response line.
        title:
          type: string
          description: >-
            A short, human-readable summary of the kind of problem. Stable
            for that kind: the same title recurs across requests and, where
            there is a `code`, across every response that sets that code.
          examples: [List failed]
        detail:
          type: string
          description: What went wrong with this request.
        instance:
          type: string
          description: The request path that produced this response.
          examples: [/v1/assets/bpmn]
        code:
          type: string
          description: >-
            The machine-readable identifier this operation's description
            names, kept for callers that already branch on it. Omitted by
            the few operations that never set one.
          examples: [LIST_FAILED]
...
    RopaPersonalDataField:
      type: object
      additionalProperties: false
      required:
        [id, ropaRecordId, formId, fieldKey, fieldLabel, dataCategory, specialCategory, sortOrder]
      properties:
        id:
          type: string
          format: uuid
        ropaRecordId:
          type: string
          format: uuid
        formId:
          type: string
        fieldKey:
          type: string
        fieldLabel:
          type: string
        dataCategory:
          type: string
        specialCategory:
          type: boolean
        sortOrder:
          type: integer
...
    RopaRecord:
      type: object
      description: >-
        A stored RoPA record, as returned by `GET /assets/ropa` and
        `GET /assets/ropa/by-bpmn-id/{bpmnProcessId}`. A superset of
        PublicRopaRecord: adds the required controllerContact and
        schemaVersion and the optional dpoContact, none of which the public
        read exposes, and status is the open three-value set rather than
        the public read's `const: active` (a query filter on the public
        read, not a difference in what the database can hold).
      additionalProperties: false
      required:
        - id
        - bpmnProcessId
        - processLevel
        - title
        - controllerName
        - controllerContact
        - purpose
        - legalBasisUri
        - legalBasisLabel
        - gdprArticle
        - dataSubjects
        - recipients
        - thirdCountryTransfers
        - retentionPeriod
        - securityMeasures
        - status
        - schemaVersion
        - personalDataFields
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        bpmnProcessId:
          type: string
        processLevel:
          type: string
          enum: [shell, subprocess]
        title:
          type: string
        controllerName:
          type: string
        controllerContact:
          type: string
        dpoContact:
          type: string
          description: Omitted when the stored value is null.
        purpose:
          type: string
        legalBasisUri:
          type: string
        legalBasisLabel:
          type: string
        gdprArticle:
          type: string
        dataSubjects:
          type: string
        recipients:
          type: string
        thirdCountryTransfers:
          type: boolean
        thirdCountryDetails:
          type: string
          description: Omitted when there are no third-country transfer details.
        retentionPeriod:
          type: string
        securityMeasures:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
        schemaVersion:
          type: integer
        personalDataFields:
          type: array
          items:
            $ref: '#/components/schemas/RopaPersonalDataField'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...